Add statuses to pawns

+++ Assets/Scripts/Pawns/IAttacker.cs
@@ -2,6 +2,6 @@
 public interface IAttacker
 {
     string name { get; }
-    float Attack { get; }
-    float Accuracy { get; }
+    int Attack { get; }
+    int Accuracy { get; }
 }
